Refactoring: GeoDataRiverSurvey::MouseEditCrosssectionCommand is moved to separated files.
