+++ libs/geodata/riversurvey/geodatariversurveycrosssectionwindowgraphicsview.cpp
@@ -5,6 +5,7 @@
 #include "geodatariversurvey.h"
 #include "geodatariversurveycrosssectionwindow.h"
 #include "geodatariversurveycrosssectionwindowgraphicsview.h"
+#include "private/geodatariversurvey_mouseeditcrosssectioncommand.h"
 
 #include <guicore/misc/qundocommandhelper.h>
 #include <guicore/project/projectdataitem.h>
@@ -594,61 +595,6 @@ void GeoDataRiverSurveyCrosssectionWindowGraphicsView::cameraZoomOutY()
 {
 	zoom(1, 1. / 1.2);
 }
-
-
-class GeoDataRiverSurvey::MouseEditCrosssectionCommand : public QUndoCommand
-{
-public:
-	MouseEditCrosssectionCommand(GeoDataRiverPathPoint* p, const GeoDataRiverCrosssection::AltitudeList& after, const GeoDataRiverCrosssection::AltitudeList& before, GeoDataRiverSurveyCrosssectionWindow* w, GeoDataRiverSurvey* rs, bool dragging = false) :
-		QUndoCommand {GeoDataRiverSurveyCrosssectionWindowGraphicsView::tr("Altitude Points Move")}
-	{
-		m_point = p;
-		m_before = before;
-		m_after = after;
-		m_window = w;
-		m_rs = rs;
-		m_dragging = dragging;
-	}
-	void redo() {
-		m_point->crosssection().AltitudeInfo() = m_after;
-		m_point->updateXSecInterpolators();
-		m_point->updateRiverShapeInterpolators();
-		if (m_dragging) {
-			m_window->updateView();
-		} else {
-			m_rs->updateShapeData();
-			m_rs->renderGraphicsView();
-			m_rs->updateCrossectionWindows();
-		}
-	}
-	void undo() {
-		m_point->crosssection().AltitudeInfo() = m_before;
-		m_point->updateXSecInterpolators();
-		m_point->updateRiverShapeInterpolators();
-		m_rs->updateShapeData();
-		m_rs->renderGraphicsView();
-		m_rs->updateCrossectionWindows();
-	}
-	int id() const {
-		return iRIC::generateCommandId("GeoDataRiverSurveyCrosssectionDragEdit");
-	}
-	bool mergeWith(const QUndoCommand* other) {
-		const MouseEditCrosssectionCommand* comm = dynamic_cast<const MouseEditCrosssectionCommand*>(other);
-		if (comm == nullptr) {return false;}
-		if (m_point != comm->m_point) {return false;}
-		if (! m_dragging) {return false;}
-		m_after = comm->m_after;
-		m_dragging = comm->m_dragging;
-		return true;
-	}
-private:
-	bool m_dragging;
-	GeoDataRiverPathPoint* m_point;
-	GeoDataRiverCrosssection::AltitudeList m_before;
-	GeoDataRiverCrosssection::AltitudeList m_after;
-	GeoDataRiverSurveyCrosssectionWindow* m_window;
-	GeoDataRiverSurvey* m_rs;
-};
 
 void GeoDataRiverSurveyCrosssectionWindowGraphicsView::mouseMoveEvent(QMouseEvent* event)
 {
