+++ libs/geodata/riversurvey/riversurvey.pro
@@ -121,6 +121,7 @@ HEADERS += gd_riversurvey_global.h \
            private/geodatariversurvey_deleteriverpathpointcommand.h \
            private/geodatariversurvey_expandcrosssectioncommand.h \
            private/geodatariversurvey_insertriverpathpointcommand.h \
+           private/geodatariversurvey_mouseeditcrosssectioncommand.h \
            private/geodatariversurvey_setdisplaysettingcommand.h \
            private/geodatariversurveycrosssectionwindow_impl.h \
            private/geodatariversurveyproxy_setsettingcommand.h
@@ -168,6 +169,7 @@ SOURCES += geodatarivercrosssection.cpp \
            private/geodatariversurvey_deleteriverpathpointcommand.cpp \
            private/geodatariversurvey_expandcrosssectioncommand.cpp \
            private/geodatariversurvey_insertriverpathpointcommand.cpp \
+           private/geodatariversurvey_mouseeditcrosssectioncommand.cpp \
            private/geodatariversurvey_setdisplaysettingcommand.cpp \
            private/geodatariversurveyproxy_setsettingcommand.cpp
 RESOURCES += riversurvey.qrc
